wsl_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyascii=0.3.2=pypi_0
anyio=4.2.0=pypi_0
asteroid-filterbanks=0.4.0=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=pypi_0
babel=2.14.0=pypi_0
bangla=0.0.2=pypi_0
blinker=1.7.0=pypi_0
blis=0.7.11=pypi_0
bnnumerizer=0.0.2=pypi_0
bnunicodenormalizer=0.1.6=pypi_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
ci-sdr=0.0.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.1=pypi_0
confection=0.1.4=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.0=pypi_0
coqpit=0.0.17=pypi_0
ctc-segmentation=1.7.4=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
cython=3.0.8=pypi_0
datasets=2.17.0=pypi_0
dateparser=1.1.8=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.8=pypi_0
distance=0.1.3=pypi_0
docopt=0.6.2=pypi_0
editdistance=0.8.1=pypi_0
einops=0.7.0=pypi_0
encodec=0.1.1=pypi_0
espnet=202402=pypi_0
espnet-model-zoo=0.1.7=pypi_0
espnet-tts-frontend=0.0.3=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fast-bss-eval=0.1.3=pypi_0
fastapi=0.109.2=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.1=pypi_0
flask=3.0.2=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.48.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
g2p-en=2.1.0=pypi_0
g2pkk=0.1.2=pypi_0
gast=0.5.4=pypi_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pypi_0
gradio=4.18.0=pypi_0
gradio-client=0.10.0=pypi_0
grpcio=1.60.1=pypi_0
gruut=2.2.3=pypi_0
gruut-ipa=0.13.0=pypi_0
gruut-lang-de=2.0.0=pypi_0
gruut-lang-en=2.0.0=pypi_0
gruut-lang-es=2.0.0=pypi_0
gruut-lang-fr=2.0.2=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
hangul-romanize=0.1.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.3=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
hyperpyyaml=1.2.2=pypi_0
idna=3.6=pypi_0
importlib-metadata=4.13.0=pypi_0
importlib-resources=6.1.1=pypi_0
inflect=7.0.0=pypi_0
inquirerpy=0.3.4=pypi_0
ipykernel=6.29.2=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.2=pypi_0
itsdangerous=2.1.2=pypi_0
jaconv=0.3.4=pypi_0
jamo=0.4.1=pypi_0
jedi=0.19.1=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonlines=1.2.0=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.7.1=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kaldiio=2.18.0=pypi_0
keras=2.15.0=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.3.0=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.9.2=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.42.0=pypi_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.2.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.7=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.10=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=2.8.8=pypi_0
nltk=3.8.1=pypi_0
num2words=0.5.13=pypi_0
numba=0.59.0=pypi_0
numpy=1.22.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
orjson=3.9.13=pypi_0
packaging=23.2=pypi_0
pandas=1.5.3=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pfzy=0.3.4=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.2.0=pypi_0
pooch=1.8.0=pypi_0
preshed=3.0.9=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.2=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=15.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydantic=2.6.1=pypi_0
pydantic-core=2.16.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pynndescent=0.5.11=pypi_0
pyparsing=3.1.1=pypi_0
pypinyin=0.44.0=pypi_0
pysbd=0.3.4=pypi_0
python=3.9.18=h955ad1f_0
python-crfsuite=0.9.10=pypi_0
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyworld=0.3.4=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.33.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.4.2=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.18.6=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
ruff=0.2.1=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.11.4=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=68.2.2=py39h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
sniffio=1.3.0=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
spacy=3.7.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
speechbrain=0.5.16=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.36.3=pypi_0
sudachidict-core=20240109=pypi_0
sudachipy=0.6.8=pypi_0
sympy=1.12=pypi_0
tensorboard=2.15.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0.post1=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
termcolor=2.4.0=pypi_0
thinc=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.0=pypi_0
torch-complex=0.4.3=pypi_0
torchaudio=2.2.0=pypi_0
torchvision=0.17.0=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
trainer=0.0.36=pypi_0
traitlets=5.14.1=pypi_0
transformers=4.37.2=pypi_0
triton=2.2.0=pypi_0
tts=0.22.0=pypi_0
typeguard=2.13.3=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023d=h04d1e81_0
tzlocal=5.2=pypi_0
umap-learn=0.5.5=pypi_0
unidecode=1.3.8=pypi_0
urllib3=2.2.0=pypi_0
uvicorn=0.27.1=pypi_0
wasabi=1.1.2=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.3.4=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.10=pypi_0
wrapt=1.14.1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
